requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
groq==0.11.0
httpx==0.27.2
python-dotenv==1.0.1
pypdf==5.1.0
scikit-learn==1.5.2
numpy==2.1.3
